merged master branch (including boatload of visual studio 64bit fixes) into c++11

Conflicts:
	CMakeLists.txt
	papara.cpp
	pars_align_gapp_seq.h
	stepwise_addition_pro.cpp
	stepwise_align.h
	tree_utils.h

+++ pars_align_seq.h
@@ -242,7 +242,7 @@ public:
         bool inL = false;
 
         if (m_tbStartA < ptrdiff_t(m_na) - 1) {
-            for (int ba = m_na - 1; ba > m_tbStartA; ba--) {
+            for (ptrdiff_t ba = m_na - 1; ba > m_tbStartA; ba--) {
                 bvec.push_back(1);
 
             }
